// frontend/src/api/models/ChangeMemberRoleDtoRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ChangeMemberRoleDtoRequest = {
    role: ChangeMemberRoleDtoRequest.role;
};
export namespace ChangeMemberRoleDtoRequest {
    export enum role {
        OWNER = 'OWNER',
        ADMIN = 'ADMIN',
        MEMBER = 'MEMBER',
    }
}